api/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Caching and Task Queue
redis==5.0.1
celery[redis]==5.3.4

# HTTP Client
requests==2.31.0
httpx==0.25.2

# Cryptography and Security
pycryptodome==3.19.0
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2023.3
pendulum==2.1.2

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Image Processing (for media files)
pillow==10.1.0

# File handling
aiofiles==23.2.1

# Environment and Configuration
python-dotenv==1.0.0

# Data Validation and Serialization
marshmallow==3.20.1
jsonschema==4.20.0

# Background Tasks
schedule==1.2.0

# Networking and WebSockets
websockets==12.0
sse-starlette==1.8.2

# Development and Testing (minimal for production)
click==8.1.7

# Performance
orjson==3.9.10